runtime gen list of plugins for analyse cmd
