+++ cmd/version/version.go
@@ -1,11 +1,7 @@
 package version
 
 import (
-	"encoding/json"
 	"fmt"
-	"log"
-	"os"
-	"path/filepath"
 
 	"github.com/spf13/cobra"
 
@@ -22,14 +18,8 @@ var (
 
 // Versions struct holds version information for the core application and plugins.
 type CoreVersions struct {
-	Versions    shared.Versions       `json:"versions"`
-	PluginsMeta map[string]PluginMeta `json:"plugins_meta"`
-}
-
-// PluginVersion holds version information for a plugin.
-type PluginMeta struct {
-	Version    string `json:"version"`
-	PluginType string `json:"plugin_type"`
+	Versions    shared.Versions              `json:"versions"`
+	PluginsMeta map[string]shared.PluginMeta `json:"plugins_meta"`
 }
 
 // Init initializes the global configuration variable.
@@ -52,45 +42,12 @@ func NewVersionCmd() *cobra.Command {
 			}
 			version := CoreVersions{
 				Versions:    versionInfo,
-				PluginsMeta: getPluginVersions(config.GetScanioPluginsHome(AppConfig)),
+				PluginsMeta: shared.GetPluginVersions(config.GetScanioPluginsHome(AppConfig), ""),
 			}
 
 			printVersionInfo(&version)
 		},
 	}
-}
-
-// readVersionFile reads and parses the version file as JSON.
-func readVersionFile(versionFilePath string) PluginMeta {
-	var pm PluginMeta
-	data, err := os.ReadFile(versionFilePath)
-	if err != nil {
-		return PluginMeta{Version: "unknown", PluginType: "unknown"}
-	}
-	if err := json.Unmarshal(data, &pm); err != nil {
-		return PluginMeta{Version: "unknown", PluginType: "unknown"}
-	}
-	return pm
-}
-
-// getPluginVersions iterates through the plugin directories and reads their version files.
-func getPluginVersions(pluginsDir string) map[string]PluginMeta {
-	pluginsMeta := make(map[string]PluginMeta)
-	entries, err := os.ReadDir(pluginsDir)
-	if err != nil {
-		log.Printf("Failed to read plugins directory: %v", err)
-		pluginsMeta["unknown"] = PluginMeta{Version: "unknown", PluginType: "unknown"}
-		return pluginsMeta
-	}
-	for _, entry := range entries {
-		if entry.IsDir() {
-			pluginName := entry.Name()
-			versionFilePath := filepath.Join(pluginsDir, pluginName, "VERSION")
-			version := readVersionFile(versionFilePath)
-			pluginsMeta[pluginName] = version
-		}
-	}
-	return pluginsMeta
 }
 
 // printVersionInfo prints the version information for the core application and plugins.
